issue-5894: service_kikimr simple refactoring - extracting different actors into different files from service.cpp
